server/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
asgiref==3.8.1
    # via django
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
beautifulsoup4==4.13.4
    # via django-bootstrap4
boto3==1.37.36
    # via einvoicing-local-ui (pyproject.toml)
botocore==1.37.36
    # via
    #   boto3
    #   s3transfer
build==1.2.2.post1
    # via pip-tools
certifi==2025.1.31
    # via requests
cffi==1.17.1
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via pip-tools
crispy-bootstrap4==2024.10
    # via einvoicing-local-ui (pyproject.toml)
cryptography==44.0.2
    # via
    #   pygstn
    #   pyopenssl
cx-oracle==8.3.0
    # via einvoicing-local-ui (pyproject.toml)
django==5.2
    # via
    #   einvoicing-local-ui (pyproject.toml)
    #   crispy-bootstrap4
    #   django-bootstrap-pagination-nigelm
    #   django-bootstrap4
    #   django-crispy-forms
    #   django-extensions
    #   django-solo
    #   django-storages
    #   jsonfield
django-bootstrap-pagination-nigelm==2.1.0
    # via einvoicing-local-ui (pyproject.toml)
django-bootstrap4==25.1
    # via einvoicing-local-ui (pyproject.toml)
django-crispy-forms==2.4
    # via
    #   einvoicing-local-ui (pyproject.toml)
    #   crispy-bootstrap4
django-environ==0.12.0
    # via einvoicing-local-ui (pyproject.toml)
django-extensions==4.1
    # via einvoicing-local-ui (pyproject.toml)
django-nose==1.4.7
    # via einvoicing-local-ui (pyproject.toml)
django-solo==2.4.0
    # via einvoicing-local-ui (pyproject.toml)
django-storages==1.14.6
    # via einvoicing-local-ui (pyproject.toml)
django-waitress==0.1.0
    # via einvoicing-local-ui (pyproject.toml)
docopt==0.6.2
    # via num2words
et-xmlfile==2.0.0
    # via openpyxl
greenlet==3.2.0
    # via sqlalchemy
gunicorn==23.0.0
    # via einvoicing-local-ui (pyproject.toml)
idna==3.10
    # via requests
inflection==0.5.1
    # via einvoicing-local-ui (pyproject.toml)
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonfield==3.1.0
    # via einvoicing-local-ui (pyproject.toml)
jsonschema==4.23.0
    # via
    #   einvoicing-local-ui (pyproject.toml)
    #   pygstn
jsonschema-specifications==2024.10.1
    # via jsonschema
ndg-httpsclient==0.5.1
    # via pusher
nose==1.3.7
    # via django-nose
num2words==0.5.14
    # via einvoicing-local-ui (pyproject.toml)
numpy==2.2.4
    # via opencv-python
opencv-python==4.11.0.86
    # via einvoicing-local-ui (pyproject.toml)
openpyxl==3.1.5
    # via einvoicing-local-ui (pyproject.toml)
packaging==24.2
    # via
    #   build
    #   gunicorn
pillow==11.2.1
    # via qrcode
pip==25.0.1
    # via pip-tools
pip-tools==7.4.1
    # via einvoicing-local-ui (pyproject.toml)
psycopg2-binary==2.9.10
    # via einvoicing-local-ui (pyproject.toml)
pusher==3.3.3
    # via einvoicing-local-ui (pyproject.toml)
pyasn1==0.6.1
    # via
    #   ndg-httpsclient
    #   pusher
pycparser==2.22
    # via cffi
pygstn @ git+https://bhartigstzen:****@bitbucket.org/suriya/pygstn.git@df24e8f6302422c14f28dc99759cef796b855df7
    # via einvoicing-local-ui (pyproject.toml)
pynacl==1.5.0
    # via pusher
pyodbc==5.2.0
    # via einvoicing-local-ui (pyproject.toml)
pyopenssl==25.0.0
    # via
    #   ndg-httpsclient
    #   pusher
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via botocore
pytz==2025.2
    # via einvoicing-local-ui (pyproject.toml)
pyyaml==6.0.2
    # via responses
qrcode==8.1
    # via einvoicing-local-ui (pyproject.toml)
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
    #   pygstn
requests==2.32.3
    # via
    #   einvoicing-local-ui (pyproject.toml)
    #   pusher
    #   pygstn
    #   responses
responses==0.25.7
    # via einvoicing-local-ui (pyproject.toml)
rpds-py==0.24.0
    # via
    #   jsonschema
    #   referencing
ruff==0.11.6
    # via einvoicing-local-ui (pyproject.toml)
s3transfer==0.11.5
    # via boto3
setuptools==78.1.0
    # via
    #   pip-tools
    #   supervisor
six==1.17.0
    # via
    #   pusher
    #   python-dateutil
soupsieve==2.6
    # via beautifulsoup4
sqlalchemy==1.4.54
    # via einvoicing-local-ui (pyproject.toml)
sqlparse==0.5.3
    # via django
supervisor==4.2.5
    # via einvoicing-local-ui (pyproject.toml)
tablib==3.8.0
    # via einvoicing-local-ui (pyproject.toml)
typing-extensions==4.13.2
    # via beautifulsoup4
urllib3==2.4.0
    # via
    #   botocore
    #   pusher
    #   requests
    #   responses
waitress==1.4.4
    # via django-waitress
wheel==0.45.1
    # via pip-tools
whitenoise==6.9.0
    # via einvoicing-local-ui (pyproject.toml)
xlsxwriter==3.2.3
    # via einvoicing-local-ui (pyproject.toml)
